# Supports parsing of huge union queries