I've found that writing the model back to well-formed query characters is very slow, so I'm tuning it.

``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22631.3296)
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.202
  [Host]     : .NET 6.0.28 (6.0.2824.12007), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.28 (6.0.2824.12007), X64 RyuJIT AVX2


```

## before
|                Method |        Mean |    Error |   StdDev |
|---------------------- |------------:|---------:|---------:|
|          SqModelParse | 1,323.82 μs | 3.668 μs | 3.252 μs |
|    SqModelCommandText |    36.03 μs | 0.144 μs | 0.134 μs |
|      CarbunqlDeepCopy |   110.70 μs | 0.280 μs | 0.233 μs |
|         CarbunqlParse | 1,357.76 μs | 2.980 μs | 2.788 μs |
| CarbunqlToOneLineText |   119.22 μs | 0.291 μs | 0.272 μs |
|        CarbunqlToText |   512.18 μs | 2.084 μs | 1.949 μs |


## after
|                Method |        Mean |    Error |   StdDev |
|---------------------- |------------:|---------:|---------:|
|          SqModelParse | 1,294.06 μs | 2.401 μs | 2.128 μs |
|    SqModelCommandText |    34.99 μs | 0.121 μs | 0.114 μs |
|      CarbunqlDeepCopy |   110.15 μs | 0.281 μs | 0.219 μs |
|         CarbunqlParse | 1,324.94 μs | 1.810 μs | 1.693 μs |
| CarbunqlToOneLineText |   114.61 μs | 0.540 μs | 0.505 μs |
|        CarbunqlToText |   468.47 μs | 1.511 μs | 1.413 μs |